bump versions, add error checks
