+++ apps/rhf-mui-demo/next.config.mjs
@@ -5,7 +5,7 @@ const nextConfig = {
     dangerouslyAllowSVG: true,
     contentDispositionType: 'attachment',
     contentSecurityPolicy: "default-src 'self'; script-src 'none'; sandbox;",
-		remotePatterns: [
+    remotePatterns: [
       {
         protocol: 'https',
         hostname: 'img.icons8.com'
