+++ apps/rhf-mui-demo/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@nish1896/rhf-mui-demo",
-  "version": "2.1.2",
+  "version": "2.1.3",
   "private": true,
   "author": "Nishant Kohli",
   "scripts": {
@@ -43,7 +43,7 @@
     "zod": "^3.23.8"
   },
   "devDependencies": {
-    "@nish1896/eslint-config": "^3.0.2",
+    "@nish1896/eslint-config": "^3.0.3",
     "@types/luxon": "^3.4.2",
     "@types/node": "^20",
     "@types/react": "^18",
