+++ apps/rhf-mui-docs/docs/customization.mdx
@@ -19,9 +19,8 @@ and [FormHelperText](https://mui.com/material-ui/api/form-helper-text/). Compone
 also include a [FormControlLabel](https://mui.com/material-ui/api/form-control-label/). 
 These can be preconfigured to apply uniform styles across all components. 
 
-Additionally, there's an option to customize the `dateAdapter` for 
-date and time pickers, with `AdapterDayjs` set as the default, following 
-[MUI's recommendation](https://mui.com/x/react-date-pickers/base-concepts/#recommended-library).
+Additionally, there's an option to customize the default `dateAdapter` for 
+date and time pickers, with `AdapterDayjs` being [recommended by Material UI](https://mui.com/x/react-date-pickers/base-concepts/#recommended-library).
 
 Starting with version 2, a new boolean flag, `allLabelsAboveFields`,
 also has been introduced to simplify label placement across form components.
