+++ apps/rhf-mui-docs/docs/introduction.mdx
@@ -35,8 +35,8 @@ Ensure that you also have installed [Material UI](https://www.npmjs.com/package/
 If not, refer to the [installation guide](https://mui.com/material-ui/getting-started/installation/). 
 For datepickers, follow the [installation guide](https://mui.com/x/react-date-pickers/getting-started/) 
 to install [@mui/x-date-pickers](https://www.npmjs.com/package/@mui/x-date-pickers).
-The form components are compatible with both [v5](https://v5.mui.com/material-ui/getting-started/) 
-and [v6](https://mui.com/material-ui/getting-started/) of material-ui.
+The form components are compatible with [v5](https://v5.mui.com/material-ui/), 
+[v6](https://v6.mui.com/material-ui/) and [v7](https://mui.com/material-ui/) of material-ui.
 
 **Optional** : If you are using a third party library for form validation, say [joi](https://www.npmjs.com/package/joi), make
 sure to install its npm package along with the [@hookform/resolvers](https://www.npmjs.com/package/@hookform/resolvers) package.
