+++ packages/rhf-mui-components/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@nish1896/rhf-mui-components",
-  "version": "2.1.2",
+  "version": "2.1.3",
   "description": "A suite of 20+ reusable Material UI components for React Hook Form to minimize your time and effort in creating and styling forms",
   "author": "Nishant Kohli",
   "homepage": "https://rhf-mui-components.netlify.app/",
@@ -14,7 +14,7 @@
   },
   "devDependencies": {
     "@eslint/js": "^9.23.0",
-    "@nish1896/eslint-flat-config": "^2.0.2",
+    "@nish1896/eslint-flat-config": "^2.0.3",
     "eslint": "^9.23.0",
     "rimraf": "^6.0.1",
     "tsc-alias": "^1.8.13"
