+++ packages/rhf-mui-components/src/utils/array.ts
@@ -1,4 +1,4 @@
-import type { StringOrNumber, StrNumObjOption, KeyValueOption } from '@/types';
+import type { StrNumObjOption } from '@/types';
 import { generateLabelValueErrMsg } from '@/utils';
 
 function isStrNumArray(arr: StrNumObjOption[]): boolean {
@@ -11,24 +11,11 @@ export function validateArray(
   labelKey?: string,
   valueKey?: string,
 ) {
+  if (!Array.isArray(options)) {
+    throw new Error(`The "options" prop of ${formElementName} must be an array.`);
+  }
   const isScalarArr = isStrNumArray(options);
   if (!isScalarArr && (!labelKey || !valueKey)) {
     throw new Error(generateLabelValueErrMsg(formElementName));
   }
 }
-
-export function isKeyValueOption(
-  option: StringOrNumber | KeyValueOption,
-  labelKey?: string,
-  valueKey?: string
-): option is KeyValueOption {
-  if (typeof option !== 'object' || option === null) {
-    return false;
-  }
-  if (!labelKey || !valueKey) {
-    return false;
-  }
-  return labelKey in option && valueKey in option;
-}
-
-
