+++ packages/rhf-mui-components/src/utils/index.ts
@@ -3,4 +3,5 @@ export * from './color';
 export * from './control';
 export * from './file';
 export * from './mui';
+export * from './object';
 export * from './text-transform';
